+++ CHANGELOG.md
@@ -7,10 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.13.0] - 2026-05-20
+
 ### Added
 
 - (Scenario) Support local context and add local variable management commands
 - (Core) Add ReadFile command and ReadFormat type for file reading 
+- (Audio) Add LoadAndPlay command to AudioCommand
 
 ## [0.12.0] - 2026-05-09